refactor(apps/web): Connection 컴포넌트 로직을 useConnection hook으로 분리
